build: Respect directory options in post install script
The `sysconfdir` or `datadir` can be customized, in which case the `post_install` script would fail.
Pass them as an argument from the Meson script.

Since the paths can be absolute and not even relative to prefix, the “correct way” [1] of using `MESON_INSTALL_DESTDIR_PREFIX` does not work.
But that should not be a problem since the option sounds like it is only distinct from `$DESTDIR/$MESON_INSTALL_PREFIX` on Windows [2], which is not a target platform.

[1]: https://mesonbuild.com/Reference-manual_builtin_meson.html#mesonadd_install_script
[2]: https://github.com/mesonbuild/meson/blob/05643006a7ee8a457456f8f0a2cbf46c395ecc25/docs/markdown/Release-notes-for-0.37.0.md?plain=1#L159

Part-of: <https://gitlab.gnome.org/GNOME/xdg-user-dirs-gtk/-/merge_requests/16>